% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{check_dropbox_access}
\alias{check_dropbox_access}
\title{`check_dropbox_access` checks if you have access to the Grattan data warehouse}
\usage{
check_dropbox_access()
}
\value{
`TRUE` if you appear to have access to the data warehouse; `FALSE` if not.
}
\description{
Use of \code{read_microdata()} relies on access to the Grattan data warehouse,
which is housed on Dropbox. The \code{check_dropbox_access()} function checks if you have 
access to the Dropbox folder.
}
