% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{get_dropbox_location}
\alias{get_dropbox_location}
\title{get_dropbox_location returns the local path to the user's Dropbox}
\usage{
get_dropbox_location(type = "business")
}
\arguments{
\item{type}{Either 'business' (the default) or 'personal'.}
}
\value{
String corresponding to the path to the user's business or personal
Dropbox.
}
\description{
get_dropbox_location returns the local path to the user's Dropbox
}
