% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_filename.R
\name{find_filename}
\alias{find_filename}
\title{find_filename finds a file in the Grattan data warehouse}
\usage{
find_filename(filename)
}
\arguments{
\item{filename}{A filename or fragment, with or without filepath,
such as "SIH15bh.dta", "SIH15bh", "ABS/SIH/2015-16/Stata/SIH15bh.dta", "SIH".
If you specify a file extension, such as ".dta", ".csv", or ".sas7bdat",
the extension must match exactly.}
}
\value{
The full local path to the file that matches 'filename' in the
Grattan data warehouse. If more than one match is found, an error will
be shown, including details of the multiple matches.
}
\description{
`find_filename()` finds files that correspond to the provided
input in the Grattan data warehouse.
}
\details{
Note that certain file extensions, including ".zip", ".txt" and
".fst" are excluded from find_filename's search. If you want to load a ".fst"
file, use `read_microdata(fast = TRUE)`
}
