% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{read_microdata loads data from the Grattan Institute data warehouse
This function requires access to the Grattan Institute data warehouse, which
is housed in the Grattan Institute Dropbox. If you do not have access to
this Dropbox the function will not work.}
\usage{
read_microdata(filename, ...)
}
\arguments{
\item{filename}{filename}

\item{...}{arguments passed to `rio::import()`. See `?rio::import`}
}
\description{
read_microdata loads data from the Grattan Institute data warehouse
This function requires access to the Grattan Institute data warehouse, which
is housed in the Grattan Institute Dropbox. If you do not have access to
this Dropbox the function will not work.
}
\examples{

\donttest{
sih_1516 <- read_microdata("SIH15bp.dta")
}

}
