% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{read_microdata loads data from the Grattan Institute data warehouse}
\usage{
read_microdata(release = NULL, provider = "ABS", date = NULL,
  format = NULL, file = NULL, path = NULL, setclass = "tbl", ...)
}
\description{
read_microdata loads data from the Grattan Institute data warehouse
}
